refactor(dom): centralize DOM utilities in utils/dom
BREAKING CHANGE: relocates all DOM utilities to a single module at utils/dom;
removes legacy dom-utils and updates all imports accordingly.

Centralizes HTML escaping, status rendering, and safe element accessors;
adds legacy unsafe helpers for migration.

Tests updated to reflect new escape sequences.
