+++ src/diff-viewer.ts
@@ -3,10 +3,10 @@
  * Provides side-by-side visualization with highlighting
  */
 
-import { escapeHtml } from "@/dom-utils";
 import { CodonLexer } from "@/lexer";
 import { compareGenomes, type MutationResult } from "@/mutations";
 import { Canvas2DRenderer } from "@/renderer";
+import { escapeHtml } from "@/utils/dom";
 import { CodonVM } from "@/vm";
 
 export interface DiffViewOptions {
