+++ src/pages/tutorial-page.ts
@@ -1,6 +1,6 @@
-import { escapeHtml, getElement } from "@/dom-utils";
 import { CodonLexer } from "@/lexer";
 import { Canvas2DRenderer } from "@/renderer";
+import { escapeHtml, getElementUnsafe as getElement } from "@/utils/dom";
 import { CodonVM } from "@/vm";
 
 // Types
